+++ app/src/androidTest/java/org/simple/clinic/di/TestAppComponent.kt
@@ -11,6 +11,8 @@ import org.simple.clinic.bloodsugar.BloodSugarRepositoryAndroidTest
 import org.simple.clinic.bp.BloodPressureHistoryListItemDataSourceTest
 import org.simple.clinic.bp.BloodPressureRepositoryAndroidTest
 import org.simple.clinic.drugs.PrescriptionRepositoryAndroidTest
+import org.simple.clinic.drugs.search.DrugRepositoryAndroidTest
+import org.simple.clinic.drugs.search.sync.DrugSyncIntegrationTest
 import org.simple.clinic.facility.FacilityRepositoryAndroidTest
 import org.simple.clinic.login.LoginUserWithOtpServerIntegrationTest
 import org.simple.clinic.medicalhistory.MedicalHistoryRepositoryAndroidTest
@@ -114,4 +116,6 @@ interface TestAppComponent {
   fun inject(target: DeleteSyncGroupDatabaseAndroidTest)
   fun inject(target: ServerRegistrationAtFacilityRule)
   fun inject(target: LookupPatientOnlineApiIntegrationTest)
+  fun inject(target: DrugRepositoryAndroidTest)
+  fun inject(target: DrugSyncIntegrationTest)
 }