+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Migrate `BloodSugarHistoryScreen` to a fragment
 - Delete patients and it's complete medical records when the retention time has passed
 - Push records to server during sync in batches
+- Implement `DrugSync`
 
 ### Features
 - [In Progress: 06 Jul 2021] Add support for finding a patient online from ID scan